Made some more adjustments
I made some more adjustments by adding a $ sign for every calculation, also added some comments for the calculations saying it is rounded to the whole number, and changed a line indicating to make sure their salary isn't a decimal value.  But overall it works fine and tested it to make sure. (p.s. commented if wanted to make a while statement to make user re-input if input is an error)