Added variables
Made the name and occupation entered to a variable because if it wasn't, the code wouldn't work, as the name and occupation entered wouldnt be able to be recognized by the future methods. Also, I tested it out with the table method, and the output of the table with the output of the calculations flows very well.