Changed up the generate csv file method
First I changed up some of the printing to csv file format as it was not organized as it was all in one line. So I wrote each section of calculations in each separate line, but I couldn't do that with the printWriter package, so i took that out and used the bufferedWriter which was much easier to write to a file. Then I added a try and catch so it automatically generates the file even if the user doesn't already have one. I added some comments and added a little text of (e.g.) to show them how it should be inputted. Finally, I called this method in the main, but I had to initialize the variables like CPP, EI, etc. so it would be known in the parameters, which were already initialized in the other methods. After these changes, the code flows perfectly as everything is calculated, formatted, and generated properly.