Do while loop in main for multiple calculations
I added a do-while loop in the main in order for multiple users to enter their information to calculate their NET pay. They have the choice to proceed or not. I also added a greet to our NET pay calculator on what it does. I also added some comments in there.