Changes in main and inputFile method
In the main, I added an option to the user whether if the user only wants to generate a table, a csv file, or a bar graph, or even all, then with that option, it will also tell the user whether the NET pay we calculated is above or below the average Canadian. I removed the duplication of the packages and I added half a docstring for the inputFile method. I added a print statement for the table method for when the table is generated. Now for the input file method, I changed it to where the user can input where the user's file is located and gave an example on how to input it, then I changed the spacing of the statements as it was too crammed together, and added minor comments throughout the code